Rename AwsLambdaAppender to StdOutAppender, plus support configuring component and environment
